requirements.txt
# XAI Suicide Detection System - Dependencies

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0
joblib>=1.1.0

# Explainable AI
lime>=0.2.0.1

# Web Application
streamlit>=1.28.0

# Data Download
kagglehub>=0.2.0

# Gemini API (optional)
google-generativeai>=0.3.0

# Additional utilities
python-dotenv>=1.0.0
